refactor: convert transfer-04 into basic-04
