+++ basic/basic-04-event-consumer/connector-with-event-subscriber/src/main/java/org/eclipse/edc/sample/extension/AssetCreatedSubscriber.java
@@ -1,5 +1,5 @@
 /*
- *  Copyright (c) 2026 Think-it GmbH
+ *  Copyright (c) 2022 Bayerische Motoren Werke Aktiengesellschaft (BMW AG)
  *
  *  This program and the accompanying materials are made available under the
  *  terms of the Apache License, Version 2.0 which is available at
@@ -8,27 +8,31 @@
  *  SPDX-License-Identifier: Apache-2.0
  *
  *  Contributors:
- *       Think-it GmbH - initial API and implementation
+ *       Bayerische Motoren Werke Aktiengesellschaft (BMW AG) - initial API and implementation
  *
  */
 
 package org.eclipse.edc.sample.extension;
 
+import org.eclipse.edc.connector.controlplane.asset.spi.event.AssetCreated;
 import org.eclipse.edc.spi.event.Event;
 import org.eclipse.edc.spi.event.EventEnvelope;
 import org.eclipse.edc.spi.event.EventSubscriber;
 import org.eclipse.edc.spi.monitor.Monitor;
 
-class TransferProcessStartedSubscriber implements EventSubscriber {
+class AssetCreatedSubscriber implements EventSubscriber {
+
     private final Monitor monitor;
 
-    TransferProcessStartedSubscriber(Monitor monitor) {
+    AssetCreatedSubscriber(Monitor monitor) {
         this.monitor = monitor;
     }
 
     @Override
-    public <E extends Event> void on(EventEnvelope<E> event) {
-        monitor.info("TransferProcessStarted event has been emitted.");
-        // do something after transfer has been started
+    public <E extends Event> void on(EventEnvelope<E> envelope) {
+        if (envelope.getPayload() instanceof AssetCreated event) {
+            monitor.info("Asset created: " + event.getAssetId());
+        }
     }
+
 }
