+++ basic/basic-04-event-consumer/connector-with-event-subscriber/src/main/java/org/eclipse/edc/sample/extension/EventSubscriberExtension.java
@@ -1,5 +1,5 @@
 /*
- *  Copyright (c) 2021 Microsoft Corporation
+ *  Copyright (c) 2022 Bayerische Motoren Werke Aktiengesellschaft (BMW AG)
  *
  *  This program and the accompanying materials are made available under the
  *  terms of the Apache License, Version 2.0 which is available at
@@ -8,28 +8,26 @@
  *  SPDX-License-Identifier: Apache-2.0
  *
  *  Contributors:
- *       Microsoft Corporation - Initial implementation
+ *       Bayerische Motoren Werke Aktiengesellschaft (BMW AG) - initial API and implementation
  *
  */
 
 package org.eclipse.edc.sample.extension;
 
-import org.eclipse.edc.connector.controlplane.transfer.spi.event.TransferProcessStarted;
+import org.eclipse.edc.connector.controlplane.asset.spi.event.AssetCreated;
 import org.eclipse.edc.runtime.metamodel.annotation.Inject;
 import org.eclipse.edc.spi.event.EventRouter;
 import org.eclipse.edc.spi.system.ServiceExtension;
 import org.eclipse.edc.spi.system.ServiceExtensionContext;
 
-public class TransferProcessStartedSubscriberExtension implements ServiceExtension {
+public class EventSubscriberExtension implements ServiceExtension {
 
     @Inject
     private EventRouter eventRouter;
 
     @Override
     public void initialize(ServiceExtensionContext context) {
-        var monitor = context.getMonitor();
-
-        eventRouter.register(TransferProcessStarted.class, new TransferProcessStartedSubscriber(monitor));
+        eventRouter.register(AssetCreated.class, new AssetCreatedSubscriber(context.getMonitor()));
     }
 
 }
