+++ basic/basic-04-event-consumer/resources/create-asset.json
@@ -1,0 +1,16 @@
+{
+  "@context": {
+    "@vocab": "https://w3id.org/edc/v0.0.1/ns/"
+  },
+  "@id": "assetId",
+  "properties": {
+    "name": "product description",
+    "contenttype": "application/json"
+  },
+  "dataAddress": {
+    "type": "HttpData",
+    "name": "Test asset",
+    "baseUrl": "https://jsonplaceholder.typicode.com/users",
+    "proxyPath": "true"
+  }
+}
